refactor(client): integrated mixins to some common shared functionalities
