+++ client/components/client/PaymentSuccessful.vue
@@ -40,9 +40,11 @@
 <script>
 
 import { mapGetters, mapMutations } from 'vuex'
+import DefaultFormMixin from '../../mixins/defaultRegistrationForm'
 
 export default {
   name: 'PaymentSuccessful',
+  mixins: [DefaultFormMixin],
   data () {
     return {
       attrs: {
@@ -97,58 +99,8 @@ export default {
     async redirectToOrders () {
       this.changeOrderPostingDone(true)
       const email = this.clientPostOrderForm.email
-      /* TODO: To make this object re-usable */
-      const defaultForm = {
-        level: 1,
-        orderPostingStep: 1,
-        email: '',
-        paperSubject: '',
-        assignmentType: '',
-        topic: '',
-        pageCount: 1,
-        sources: 0,
-        wordCount: 1,
-        deadlineDate: '',
-        deadlineTime: '',
-        studyLevel: '',
-        citationStyleId: '',
-        serviceType: 'Writing',
-        supportingFiles: [],
-        instructions: '',
-        orderId: 0,
-        writerBids: [],
-        selectedWriter: {
-          name: '',
-          id: ''
-        },
-        paymentSummary: {
-          cpp: 0,
-          currencyCode: '',
-          paperPrice: 0,
-          leastPaperPrice: 0,
-          totalPrice: 0,
-          discount: 0,
-          extrasList: [],
-          extrasTotalPrice: 0
-        },
-        orderSavingProgress: {
-          details: false,
-          payment: false
-        },
-        stepStatus: {
-          step1: true,
-          step2: true,
-          step3: true,
-          step4: true
-        },
-        addFunds: {
-          paymentSuccessful: false,
-          paymentFailed: false,
-          notYetPaid: true
-        }
-      }
       /* Resetting the registration state */
-      this.changeWholeClientPostOrderForm(defaultForm)
+      this.changeWholeClientPostOrderForm(this.defaultForm)
       /* Before setting the client email. This is important because we need it when the client wants to post another order */
       this.changeClientPostOrderForm({
         key: 'email',
